refactor : WinningLotto 리팩토링 이름 변경 getter를 쓰지 않고 메시지를 보내는 형태로 리팩토링
