+++ src/main/java/lotto/domain/LottoNumber.java
@@ -14,17 +14,8 @@ public class LottoNumber {
         if (this.numbers == null) throw new IllegalArgumentException();
     }
 
-    public int getMatchCount(WinningNumber winningNumber) {
-        List<Integer> winningNumbers = winningNumber.getNumbers();
-        return (int) winningNumbers.stream()
-                .filter(this.numbers::contains)
-                .count();
-    }
-
-    public boolean isBonusContains(WinningNumber winningNumber) {
-        Integer bonusNumber = winningNumber.getBonusNumber();
-        return this.numbers.stream()
-                .anyMatch(bonusNumber::equals);
+    public boolean contains(Integer number){
+        return this.numbers.contains(number);
     }
 
     @Override
